Add support for custom highlighter.js wrapper
On the markdown-it page it explains how to use a custom wrapper for
highlighter.js. The markdown-it plugin for Docpad did not yet have a way
to configure this.

This commit adds a new configuration option name highligherWrapper. When
this is a string and useHighlighter is true than the result from
highlighter.js will be inserted into the custom wrapper. The custom
wrapper should contain the string `{{code}}`, this will replaced with
the output from highlighter.js.